create util for email otp widget
