+++ packages/scaffold/index.ts
@@ -23,7 +23,6 @@ export * from './src/views/w3m-email-verify-device-view/index.js'
 export * from './src/views/w3m-approve-transaction-view/index.js'
 export * from './src/views/w3m-upgrade-wallet-view/index.js'
 export * from './src/views/w3m-update-email-wallet-view/index.js'
-export * from './src/views/w3m-update-email-wallet-waiting-view/index.js'
 
 export * from './src/partials/w3m-all-wallets-list/index.js'
 export * from './src/partials/w3m-all-wallets-search/index.js'
