+++ packages/scaffold/src/utils/w3m-email-otp-widget/styles.ts
@@ -1,0 +1,7 @@
+import { css } from 'lit'
+
+export default css`
+  wui-loading-spinner {
+    margin: 9px auto;
+  }
+`
