+++ packages/scaffold/src/views/w3m-update-email-wallet-view/index.ts
@@ -86,7 +86,6 @@ export class W3mUpdateEmailWalletView extends LitElement {
 
       await emailConnector.provider.updateEmail({ email: this.email })
       EventsController.sendEvent({ type: 'track', event: 'EMAIL_EDIT' })
-      RouterController.replace('UpdateEmailWalletWaiting', { email: this.email })
     } catch (error) {
       SnackController.showError(error)
       this.loading = false
