--- packages/scaffold/src/views/w3m-update-email-wallet-waiting-view/styles.ts
@@ -1,8 +1,0 @@
-import { css } from 'lit'
-
-export default css`
-  wui-icon-box {
-    height: var(--wui-icon-box-size-xl);
-    width: var(--wui-icon-box-size-xl);
-  }
-`
